Updated templates and styles
