+++ app/components/sidebar.js
@@ -6,25 +6,7 @@ import { inject as service } from '@ember/service';
 export default class SidebarComponent extends Component {
   @service country;
 
-  @tracked sidebarWidth = 0;
-
   @tracked searchText = '';
-
-  countryId = (name) => name.split(' ').join('').toLowerCase();
-
-  constructor() {
-    super(...arguments);
-
-    const selectedCountry = document.querySelector(
-      `#${this.countryId(this.country.selectedCountry)}`
-    );
-
-    if (selectedCountry !== null) {
-      setTimeout(() => {
-        selectedCountry.scrollIntoView();
-      }, 0);
-    }
-  }
 
   get countryList() {
     return Object.values(this.country.availableCountries)
@@ -42,19 +24,9 @@ export default class SidebarComponent extends Component {
   }
 
   @action
-  toggleNav() {
-    this.searchText = '';
-    if (this.sidebarWidth !== 0) {
-      this.sidebarWidth = 0;
-      // const mainBody = document.querySelector('#main');
-      // mainBody.style.backgroundColor = 'rgba(0,0,0,0.4)';
-    } else this.sidebarWidth = 250;
-  }
-
-  @action
   selectCountry(newCountry) {
     this.country.changeCountry(newCountry);
-    this.toggleNav();
+    this.args.toggleNav();
     this.searchText = '';
   }
