+++ app/serializers/currency.js
@@ -2,6 +2,7 @@ import RESTSerializer from '@ember-data/serializer/rest';
 import currencySymbols from 'ember-app/utils/symbols';
 import { countries as countriesList } from 'ember-app/utils/countries';
 
+/**
 const getCountryFlag = (currency) => {
   const country = countriesList.filter(
     ({ currencyCode }) => currencyCode === currency
@@ -16,6 +17,7 @@ const getCountryFlag = (currency) => {
     else if (img.onerror) return '';
   }
 };
+ */
 
 export default class CurrencySerializer extends RESTSerializer {
   normalizeResponse(store, model, payload, ...rest) {
