+++ app/styles/components/_sidebar.scss
@@ -41,7 +41,7 @@
   background-blend-mode: multiply;
 
   @include themify($themes) {
-    background-image: themed('invertedMainBackground');
+    background-image: themed('mainBackground');
   }
 
   .sidebar_top {
@@ -56,7 +56,7 @@
     z-index: 200;
 
     @include themify($themes) {
-      background-image: themed('invertedMainBackground');
+      background-image: themed('mainBackground');
     }
 
     .label {
@@ -85,7 +85,7 @@
       margin-left: 5px;
 
       @include themify($themes) {
-        color: themed('background');
+        color: themed('sidebarText');
       }
     }
   }
@@ -122,7 +122,7 @@
     font-weight: 700;
 
     @include themify($themes) {
-      color: themed('background');
+      color: themed('sidebarText');
     }
   }
 }
