+++ app/templates/index.hbs
@@ -2,8 +2,12 @@
 
 <div class='theme-{{this.theme}}'>
   <nav class='navbar sticky-top justify-content-between'>
-    <div class='container-fluid navbar-content'>
-      <Sidebar />
+    <div class='container-fluid navbar-content' {{on "keydown" this.onKeyPress}}>
+      <Sidebar
+        @countryId={{this.countryId}}
+        @sidebarWidth={{this.sidebarWidth}}
+        @toggleNav={{this.toggleNav}}
+      />
       <img src='assets/images/logo.png' alt='Currency' class='logo' />
       <div class='d-flex justify-content-center align-items-center'>
         <label class='switch'>
@@ -19,7 +23,7 @@
     </div>
   </nav>
 
-  <div class='main'>
+  <div class='main' {{on "click" this.onClickWhenSidebarOpen}}>
     <div class='container py-3'>
       <div class='row mb-4'>
         <div class='col'>
